refactor(WinningLotto): LottoTest에 있는 당첨숫자 가져오는 메소드를 WinningLotto로 이동
